requirements.txt
aioredis==1.2.0
asgiref==3.1.2
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
daphne==2.3.0
Django==2.1.7
django-crispy-forms==1.7.2
django-widget-tweaks
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
msgpack==0.6.1
pillow==6.0
PyHamcrest==1.9.0
pylint==2.3.1
python-decouple
pytz==2018.9
selenium
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
txaio==18.8.1
typed-ast==1.3.1
wrapt==1.11.1
zope.interface==4.6.0

